+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Enhancements
 * Adds support for `NotificationConfigurationSubscribableChoice` type `Project` to support Project Level Notifications by @jillirami [#1350](https://github.com/hashicorp/go-tfe/pull/1350)
+* Adds `Explorer` to support querying organization data across workspaces and exporting it as CSV by @a9logic [#PR](https://github.com/hashicorp/go-tfe/pull/PR)
 
 # v1.108.0